requirements.txt
matplotlib==2.2.3
numpy==1.21.5
packaging==20.9
pandas==1.1.5
pytorch-lightning==1.5.8
PyYAML==5.4.1
regex==2021.11.10
scikit-learn==1.0.2
scipy==1.1.0
seaborn==0.11.2
timm==0.4.12
tokenizers==0.10.1
torch==1.9.0
torchaudio==0.9.0
torchmetrics==0.6.2
torchtext==0.6.0
torchvision==0.10.0
tqdm==4.62.3
transformers==4.15.0
huggingface-hub==0.2.1
luigi==3.0.1
pydantic==1.9.0
icecream==2.1.1
executing==0.3.1
wandb==0.12.9
scikit-image==0.14.2
opencv-python==3.4.2.17
h5py==3.1.0
Pillow==9.1.1
einops==0.4.1